add some csvs

+++ csv_import.py
@@ -14,5 +14,5 @@ class csv_import:
                     rank = int(csv_surfer["Rank"])
                 else:
                     rank = -1
-                surfers.append(surfer.Surfer(csv_surfer["Surfer"], int(csv_surfer["Odds"]), csv_surfer["Tier"], csv_surfer["Heat"], rank=rank))
+                surfers.append(surfer.Surfer(csv_surfer["Surfer"], float(csv_surfer["Odds"]), csv_surfer["Tier"], csv_surfer["Heat"], rank=rank))
         return surfer.SurferPool(surfers)
